+++ Makefile
@@ -119,7 +119,6 @@ COMMANDS += txpool
 COMMANDS += verkle
 COMMANDS += evm
 COMMANDS += sentinel
-COMMANDS += erigon-el
 COMMANDS += caplin-phase1
 COMMANDS += caplin-regression
