+++ cl/phase1/execution_client/execution_client.go
@@ -2,27 +2,102 @@ package execution_client
 
 import (
 	"context"
+	"encoding/binary"
 	"fmt"
+	"math/big"
 	"time"
 
 	"github.com/c2h5oh/datasize"
+	"github.com/holiman/uint256"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces/execution"
+	types2 "github.com/ledgerwatch/erigon-lib/gointerfaces/types"
+
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials/insecure"
 	"google.golang.org/grpc/keepalive"
 
 	"github.com/ledgerwatch/erigon/cl/cltypes"
-	"github.com/ledgerwatch/erigon/cmd/erigon-el/eth1"
 	"github.com/ledgerwatch/erigon/core/types"
-	"github.com/ledgerwatch/erigon/ethdb/privateapi"
+	"github.com/ledgerwatch/erigon/turbo/engineapi"
 )
 
 // ExecutionClient interfaces with the Erigon-EL component consensus side.
 type ExecutionClient struct {
 	client execution.ExecutionClient
 	ctx    context.Context
+}
+
+func HeaderRpcToHeader(header *execution.Header) (*types.Header, error) {
+	var blockNonce types.BlockNonce
+	binary.BigEndian.PutUint64(blockNonce[:], header.Nonce)
+	h := &types.Header{
+		ParentHash:  gointerfaces.ConvertH256ToHash(header.ParentHash),
+		UncleHash:   gointerfaces.ConvertH256ToHash(header.OmmerHash),
+		Coinbase:    gointerfaces.ConvertH160toAddress(header.Coinbase),
+		Root:        gointerfaces.ConvertH256ToHash(header.StateRoot),
+		TxHash:      gointerfaces.ConvertH256ToHash(header.TransactionHash),
+		ReceiptHash: gointerfaces.ConvertH256ToHash(header.ReceiptRoot),
+		Bloom:       gointerfaces.ConvertH2048ToBloom(header.LogsBloom),
+		Difficulty:  gointerfaces.ConvertH256ToUint256Int(header.Difficulty).ToBig(),
+		Number:      big.NewInt(int64(header.BlockNumber)),
+		GasLimit:    header.GasLimit,
+		GasUsed:     header.GasUsed,
+		Time:        header.Timestamp,
+		Extra:       header.ExtraData,
+		MixDigest:   gointerfaces.ConvertH256ToHash(header.PrevRandao),
+		Nonce:       blockNonce,
+	}
+	if header.BaseFeePerGas != nil {
+		h.BaseFee = gointerfaces.ConvertH256ToUint256Int(header.BaseFeePerGas).ToBig()
+	}
+	if header.WithdrawalHash != nil {
+		h.WithdrawalsHash = new(libcommon.Hash)
+		*h.WithdrawalsHash = gointerfaces.ConvertH256ToHash(header.WithdrawalHash)
+	}
+	blockHash := gointerfaces.ConvertH256ToHash(header.BlockHash)
+	if blockHash != h.Hash() {
+		return nil, fmt.Errorf("block %d, %x has invalid hash. expected: %x", header.BlockNumber, h.Hash(), blockHash)
+	}
+	return h, nil
+}
+
+func HeaderToHeaderRPC(header *types.Header) *execution.Header {
+	difficulty := new(uint256.Int)
+	difficulty.SetFromBig(header.Difficulty)
+
+	var baseFeeReply *types2.H256
+	if header.BaseFee != nil {
+		var baseFee uint256.Int
+		baseFee.SetFromBig(header.BaseFee)
+		baseFeeReply = gointerfaces.ConvertUint256IntToH256(&baseFee)
+	}
+	var withdrawalHashReply *types2.H256
+	if header.WithdrawalsHash != nil {
+		withdrawalHashReply = gointerfaces.ConvertHashToH256(*header.WithdrawalsHash)
+	}
+	return &execution.Header{
+		ParentHash:      gointerfaces.ConvertHashToH256(header.ParentHash),
+		Coinbase:        gointerfaces.ConvertAddressToH160(header.Coinbase),
+		StateRoot:       gointerfaces.ConvertHashToH256(header.Root),
+		TransactionHash: gointerfaces.ConvertHashToH256(header.TxHash),
+		LogsBloom:       gointerfaces.ConvertBytesToH2048(header.Bloom[:]),
+		ReceiptRoot:     gointerfaces.ConvertHashToH256(header.ReceiptHash),
+		PrevRandao:      gointerfaces.ConvertHashToH256(header.MixDigest),
+		BlockNumber:     header.Number.Uint64(),
+		Nonce:           header.Nonce.Uint64(),
+		GasLimit:        header.GasLimit,
+		GasUsed:         header.GasUsed,
+		Timestamp:       header.Time,
+		ExtraData:       header.Extra,
+		Difficulty:      gointerfaces.ConvertUint256IntToH256(difficulty),
+		BlockHash:       gointerfaces.ConvertHashToH256(header.Hash()),
+		OmmerHash:       gointerfaces.ConvertHashToH256(header.UncleHash),
+		BaseFeePerGas:   baseFeeReply,
+		WithdrawalHash:  withdrawalHashReply,
+	}
+
 }
 
 // NewExecutionClient establishes a client-side connection with Erigon-EL
@@ -59,7 +134,7 @@ func NewExecutionClient(ctx context.Context, addr string) (*ExecutionClient, err
 func (ec *ExecutionClient) InsertHeaders(headers []*types.Header) error {
 	grpcHeaders := make([]*execution.Header, 0, len(headers))
 	for _, header := range headers {
-		grpcHeaders = append(grpcHeaders, eth1.HeaderToHeaderRPC(header))
+		grpcHeaders = append(grpcHeaders, HeaderToHeaderRPC(header))
 	}
 	_, err := ec.client.InsertHeaders(ec.ctx, &execution.InsertHeadersRequest{Headers: grpcHeaders})
 	return err
@@ -76,7 +151,7 @@ func (ec *ExecutionClient) InsertBodies(bodies []*types.RawBody, blockHashes []l
 			BlockHash:    gointerfaces.ConvertHashToH256(blockHashes[i]),
 			BlockNumber:  blockNumbers[i],
 			Transactions: body.Transactions,
-			Withdrawals:  privateapi.ConvertWithdrawalsToRpc(body.Withdrawals),
+			Withdrawals:  engineapi.ConvertWithdrawalsToRpc(body.Withdrawals),
 		})
 	}
 	_, err := ec.client.InsertBodies(ec.ctx, &execution.InsertBodiesRequest{Bodies: grpcBodies})
@@ -128,7 +203,7 @@ func (ec *ExecutionClient) ReadHeader(number uint64, blockHash libcommon.Hash) (
 		return nil, err
 	}
 
-	return eth1.HeaderRpcToHeader(resp.Header)
+	return HeaderRpcToHeader(resp.Header)
 }
 
 func (ec *ExecutionClient) ReadExecutionPayload(number uint64, blockHash libcommon.Hash) (*cltypes.Eth1Block, error) {
@@ -153,7 +228,7 @@ func (ec *ExecutionClient) ReadBody(number uint64, blockHash libcommon.Hash) (*t
 	}
 	uncles := make([]*types.Header, 0, len(resp.Body.Uncles))
 	for _, uncle := range resp.Body.Uncles {
-		h, err := eth1.HeaderRpcToHeader(uncle)
+		h, err := HeaderRpcToHeader(uncle)
 		if err != nil {
 			return nil, err
 		}
@@ -162,6 +237,6 @@ func (ec *ExecutionClient) ReadBody(number uint64, blockHash libcommon.Hash) (*t
 	return &types.RawBody{
 		Transactions: resp.Body.Transactions,
 		Uncles:       uncles,
-		Withdrawals:  privateapi.ConvertWithdrawalsFromRpc(resp.Body.Withdrawals),
+		Withdrawals:  engineapi.ConvertWithdrawalsFromRpc(resp.Body.Withdrawals),
 	}, nil
 }
