+++ cl/phase1/execution_client/execution_engine.go
@@ -8,13 +8,13 @@ import (
 	"github.com/holiman/uint256"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces"
-	"github.com/ledgerwatch/erigon-lib/gointerfaces/remote"
+	"github.com/ledgerwatch/erigon-lib/gointerfaces/engine"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces/types"
 	types2 "github.com/ledgerwatch/erigon/core/types"
+	"github.com/ledgerwatch/erigon/turbo/engineapi"
 
 	"github.com/ledgerwatch/erigon/cl/clparams"
 	"github.com/ledgerwatch/erigon/cl/cltypes"
-	"github.com/ledgerwatch/erigon/ethdb/privateapi"
 )
 
 var errContextExceeded = "rpc error: code = DeadlineExceeded desc = context deadline exceeded"
@@ -29,25 +29,24 @@ type ExecutionEngine interface {
 // ExecutionEnginePhase1 is just the normal engine api.
 type ExecutionEnginePhase1 struct {
 	// Either execution server or client
-	executionServer remote.ETHBACKENDServer
-	executionClient remote.ETHBACKENDClient
+	engineClient engine.EngineClient
 
 	ctx context.Context
 }
 
 // NewExecutionEnginePhase1FromServer use ethbackend server
-func NewExecutionEnginePhase1FromServer(ctx context.Context, executionServer remote.ETHBACKENDServer) *ExecutionEnginePhase1 {
+func NewExecutionEnginePhase1FromServer(ctx context.Context, engineClient engine.EngineClient) *ExecutionEnginePhase1 {
 	return &ExecutionEnginePhase1{
-		executionServer: executionServer,
-		ctx:             ctx,
+		engineClient: engineClient,
+		ctx:          ctx,
 	}
 }
 
 // NewExecutionEnginePhase1FromServer use ethbackend client
-func NewExecutionEnginePhase1FromClient(ctx context.Context, executionClient remote.ETHBACKENDClient) *ExecutionEnginePhase1 {
+func NewExecutionEnginePhase1FromClient(ctx context.Context, engineClient engine.EngineClient) *ExecutionEnginePhase1 {
 	return &ExecutionEnginePhase1{
-		executionClient: executionClient,
-		ctx:             ctx,
+		engineClient: engineClient,
+		ctx:          ctx,
 	}
 }
 
@@ -59,12 +58,9 @@ func (e *ExecutionEnginePhase1) NewPayload(payload *cltypes.Eth1Block) error {
 	ctx, cancel := context.WithTimeout(e.ctx, 3*time.Second)
 	defer cancel()
 
-	var status *remote.EnginePayloadStatus
-	if e.executionServer != nil {
-		status, err = e.executionServer.EngineNewPayload(ctx, grpcMessage)
-	} else if e.executionClient != nil {
-		status, err = e.executionClient.EngineNewPayload(ctx, grpcMessage)
-	}
+	var status *engine.EnginePayloadStatus
+
+	status, err = e.engineClient.EngineNewPayload(ctx, grpcMessage)
 	// Ignore timeouts
 	if err != nil {
 		if err.Error() == errContextExceeded {
@@ -72,18 +68,18 @@ func (e *ExecutionEnginePhase1) NewPayload(payload *cltypes.Eth1Block) error {
 		}
 		return err
 	}
-	if status.Status == remote.EngineStatus_INVALID {
+	if status.Status == engine.EngineStatus_INVALID {
 		return fmt.Errorf("invalid block")
 	}
-	if status.Status == remote.EngineStatus_INVALID_BLOCK_HASH {
+	if status.Status == engine.EngineStatus_INVALID_BLOCK_HASH {
 		return fmt.Errorf("invalid block hash")
 	}
 	return err
 }
 
 func (e *ExecutionEnginePhase1) ForkChoiceUpdate(finalized libcommon.Hash, head libcommon.Hash) error {
-	grpcMessage := &remote.EngineForkChoiceUpdatedRequest{
-		ForkchoiceState: &remote.EngineForkChoiceState{
+	grpcMessage := &engine.EngineForkChoiceUpdatedRequest{
+		ForkchoiceState: &engine.EngineForkChoiceState{
 			HeadBlockHash:      gointerfaces.ConvertHashToH256(head),
 			SafeBlockHash:      gointerfaces.ConvertHashToH256(head),
 			FinalizedBlockHash: gointerfaces.ConvertHashToH256(finalized),
@@ -92,11 +88,8 @@ func (e *ExecutionEnginePhase1) ForkChoiceUpdate(finalized libcommon.Hash, head 
 	var err error
 	ctx, cancel := context.WithTimeout(e.ctx, 3*time.Second)
 	defer cancel()
-	if e.executionClient != nil {
-		_, err = e.executionClient.EngineForkChoiceUpdated(ctx, grpcMessage)
-	} else if e.executionServer != nil {
-		_, err = e.executionServer.EngineForkChoiceUpdated(ctx, grpcMessage)
-	}
+
+	_, err = e.engineClient.EngineForkChoiceUpdated(ctx, grpcMessage)
 	// Ignore timeouts
 	if err != nil && err.Error() == errContextExceeded {
 		return nil
@@ -145,11 +138,11 @@ func convertPayloadToGrpc(e *cltypes.Eth1Block) (*types.ExecutionPayload, error)
 		BaseFeePerGas: gointerfaces.ConvertUint256IntToH256(baseFee),
 		BlockHash:     gointerfaces.ConvertHashToH256(e.BlockHash),
 		Transactions:  e.Transactions.UnderlyngReference(),
-		Withdrawals:   privateapi.ConvertWithdrawalsToRpc(withdrawals),
+		Withdrawals:   engineapi.ConvertWithdrawalsToRpc(withdrawals),
 	}
 	if e.Version() >= clparams.CapellaVersion {
 		res.Version = 2
-		res.Withdrawals = privateapi.ConvertWithdrawalsToRpc(withdrawals)
+		res.Withdrawals = engineapi.ConvertWithdrawalsToRpc(withdrawals)
 	}
 
 	if e.Version() >= clparams.DenebVersion {
