+++ cmd/rpcdaemon/commands/corner_cases_support_test.go
@@ -18,7 +18,7 @@ func TestNotFoundMustReturnNil(t *testing.T) {
 	require := require.New(t)
 	m, _, _ := rpcdaemontest.CreateTestSentry(t)
 	api := NewEthAPI(newBaseApiForTest(m),
-		m.DB, nil, nil, nil, 5000000, 100_000, log.New())
+		m.DB, nil, nil, nil, nil, 5000000, 100_000, log.New())
 	ctx := context.Background()
 
 	a, err := api.GetTransactionByBlockNumberAndIndex(ctx, 10_000, 1)
