+++ turbo/engineapi/engine_server_test.go
@@ -1,4 +1,4 @@
-package privateapi
+package engineapi
 
 import (
 	"context"
@@ -7,15 +7,14 @@ import (
 	"github.com/ledgerwatch/erigon-lib/chain"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces"
-	"github.com/ledgerwatch/erigon-lib/gointerfaces/remote"
+	"github.com/ledgerwatch/erigon-lib/gointerfaces/engine"
 	types2 "github.com/ledgerwatch/erigon-lib/gointerfaces/types"
 	"github.com/ledgerwatch/erigon-lib/kv"
 	"github.com/ledgerwatch/erigon-lib/kv/memdb"
 	"github.com/stretchr/testify/require"
 
 	"github.com/ledgerwatch/erigon/core/rawdb"
-	"github.com/ledgerwatch/erigon/turbo/engineapi"
-	"github.com/ledgerwatch/erigon/turbo/shards"
+	"github.com/ledgerwatch/erigon/turbo/engineapi/engine_helpers"
 	"github.com/ledgerwatch/erigon/turbo/stages/headerdownload"
 	"github.com/ledgerwatch/log/v3"
 )
@@ -96,11 +95,10 @@ func TestMockDownloadRequest(t *testing.T) {
 	makeTestDb(ctx, db)
 	hd := headerdownload.NewHeaderDownload(0, 0, nil, nil, logger)
 	hd.SetPOSSync(true)
-	events := shards.NewEvents()
-	backend := NewEthBackendServer(ctx, nil, db, events, nil, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, hd, false, logger)
+	backend := NewEngineServer(ctx, logger, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, db, nil, hd, false)
 
 	var err error
-	var reply *remote.EnginePayloadStatus
+	var reply *engine.EnginePayloadStatus
 	done := make(chan bool)
 
 	go func() {
@@ -109,10 +107,10 @@ func TestMockDownloadRequest(t *testing.T) {
 	}()
 
 	hd.BeaconRequestList.WaitForRequest(true, false)
-	hd.PayloadStatusCh <- engineapi.PayloadStatus{Status: remote.EngineStatus_SYNCING}
+	hd.PayloadStatusCh <- engine_helpers.PayloadStatus{Status: engine.EngineStatus_SYNCING}
 	<-done
 	require.NoError(err)
-	require.Equal(reply.Status, remote.EngineStatus_SYNCING)
+	require.Equal(reply.Status, engine.EngineStatus_SYNCING)
 	require.Nil(reply.LatestValidHash)
 
 	// If we get another request we don't need to process it with processDownloadCh and ignore it and return Syncing status
@@ -124,7 +122,7 @@ func TestMockDownloadRequest(t *testing.T) {
 	<-done
 	// Same result as before
 	require.NoError(err)
-	require.Equal(reply.Status, remote.EngineStatus_SYNCING)
+	require.Equal(reply.Status, engine.EngineStatus_SYNCING)
 	require.Nil(reply.LatestValidHash)
 
 	// However if we simulate that we finish reverse downloading the chain by updating the head, we just execute 1:1
@@ -141,7 +139,7 @@ func TestMockDownloadRequest(t *testing.T) {
 	<-done
 
 	require.NoError(err)
-	require.Equal(reply.Status, remote.EngineStatus_SYNCING)
+	require.Equal(reply.Status, engine.EngineStatus_SYNCING)
 	require.Nil(reply.LatestValidHash)
 }
 
@@ -156,11 +154,10 @@ func TestMockValidExecution(t *testing.T) {
 	hd := headerdownload.NewHeaderDownload(0, 0, nil, nil, logger)
 	hd.SetPOSSync(true)
 
-	events := shards.NewEvents()
-	backend := NewEthBackendServer(ctx, nil, db, events, nil, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, hd, false, logger)
+	backend := NewEngineServer(ctx, logger, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, db, nil, hd, false)
 
 	var err error
-	var reply *remote.EnginePayloadStatus
+	var reply *engine.EnginePayloadStatus
 	done := make(chan bool)
 
 	go func() {
@@ -170,14 +167,14 @@ func TestMockValidExecution(t *testing.T) {
 
 	hd.BeaconRequestList.WaitForRequest(true, false)
 
-	hd.PayloadStatusCh <- engineapi.PayloadStatus{
-		Status:          remote.EngineStatus_VALID,
+	hd.PayloadStatusCh <- engine_helpers.PayloadStatus{
+		Status:          engine.EngineStatus_VALID,
 		LatestValidHash: payload3Hash,
 	}
 	<-done
 
 	require.NoError(err)
-	require.Equal(reply.Status, remote.EngineStatus_VALID)
+	require.Equal(reply.Status, engine.EngineStatus_VALID)
 	replyHash := gointerfaces.ConvertH256ToHash(reply.LatestValidHash)
 	require.Equal(replyHash[:], payload3Hash[:])
 }
@@ -193,11 +190,10 @@ func TestMockInvalidExecution(t *testing.T) {
 	hd := headerdownload.NewHeaderDownload(0, 0, nil, nil, logger)
 	hd.SetPOSSync(true)
 
-	events := shards.NewEvents()
-	backend := NewEthBackendServer(ctx, nil, db, events, nil, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, hd, false, logger)
+	backend := NewEngineServer(ctx, logger, &chain.Config{TerminalTotalDifficulty: libcommon.Big1}, nil, db, nil, hd, false)
 
 	var err error
-	var reply *remote.EnginePayloadStatus
+	var reply *engine.EnginePayloadStatus
 	done := make(chan bool)
 
 	go func() {
@@ -207,14 +203,14 @@ func TestMockInvalidExecution(t *testing.T) {
 
 	hd.BeaconRequestList.WaitForRequest(true, false)
 	// Simulate invalid status
-	hd.PayloadStatusCh <- engineapi.PayloadStatus{
-		Status:          remote.EngineStatus_INVALID,
+	hd.PayloadStatusCh <- engine_helpers.PayloadStatus{
+		Status:          engine.EngineStatus_INVALID,
 		LatestValidHash: startingHeadHash,
 	}
 	<-done
 
 	require.NoError(err)
-	require.Equal(reply.Status, remote.EngineStatus_INVALID)
+	require.Equal(reply.Status, engine.EngineStatus_INVALID)
 	replyHash := gointerfaces.ConvertH256ToHash(reply.LatestValidHash)
 	require.Equal(replyHash[:], startingHeadHash[:])
 }
@@ -229,8 +225,7 @@ func TestNoTTD(t *testing.T) {
 
 	hd := headerdownload.NewHeaderDownload(0, 0, nil, nil, logger)
 
-	events := shards.NewEvents()
-	backend := NewEthBackendServer(ctx, nil, db, events, nil, &chain.Config{}, nil, hd, false, logger)
+	backend := NewEngineServer(ctx, logger, &chain.Config{}, nil, db, nil, hd, false)
 
 	var err error
