+++ turbo/stages/sentry_mock_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/holiman/uint256"
 	libcommon "github.com/ledgerwatch/erigon-lib/common"
-	"github.com/ledgerwatch/erigon-lib/gointerfaces/remote"
+	"github.com/ledgerwatch/erigon-lib/gointerfaces/engine"
 	"github.com/ledgerwatch/erigon-lib/gointerfaces/sentry"
 	"github.com/ledgerwatch/erigon/core/rawdb"
 	"github.com/ledgerwatch/log/v3"
@@ -19,7 +19,7 @@ import (
 	"github.com/ledgerwatch/erigon/eth/protocols/eth"
 	"github.com/ledgerwatch/erigon/params"
 	"github.com/ledgerwatch/erigon/rlp"
-	"github.com/ledgerwatch/erigon/turbo/engineapi"
+	"github.com/ledgerwatch/erigon/turbo/engineapi/engine_helpers"
 	"github.com/ledgerwatch/erigon/turbo/stages"
 )
 
@@ -508,7 +508,7 @@ func TestForkchoiceToGenesis(t *testing.T) {
 	m := stages.MockWithZeroTTD(t, false)
 
 	// Trivial forkChoice: everything points to genesis
-	forkChoiceMessage := engineapi.ForkChoiceMessage{
+	forkChoiceMessage := engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      m.Genesis.Hash(),
 		SafeBlockHash:      m.Genesis.Hash(),
 		FinalizedBlockHash: m.Genesis.Hash(),
@@ -526,14 +526,14 @@ func TestForkchoiceToGenesis(t *testing.T) {
 	assert.Equal(t, m.Genesis.Hash(), rawdb.ReadHeadBlockHash(tx))
 
 	payloadStatus := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus.Status)
 }
 
 func TestBogusForkchoice(t *testing.T) {
 	m := stages.MockWithZeroTTD(t, false)
 
 	// Bogus forkChoice: head points to rubbish
-	forkChoiceMessage := engineapi.ForkChoiceMessage{
+	forkChoiceMessage := engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      libcommon.HexToHash("11111111111111111111"),
 		SafeBlockHash:      m.Genesis.Hash(),
 		FinalizedBlockHash: m.Genesis.Hash(),
@@ -549,10 +549,10 @@ func TestBogusForkchoice(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	payloadStatus := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_SYNCING, payloadStatus.Status)
+	assert.Equal(t, engine.EngineStatus_SYNCING, payloadStatus.Status)
 
 	// Now send a correct forkChoice
-	forkChoiceMessage = engineapi.ForkChoiceMessage{
+	forkChoiceMessage = engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      m.Genesis.Hash(),
 		SafeBlockHash:      m.Genesis.Hash(),
 		FinalizedBlockHash: m.Genesis.Hash(),
@@ -564,7 +564,7 @@ func TestBogusForkchoice(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	payloadStatus = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus.Status)
 }
 
 func TestPoSDownloader(t *testing.T) {
@@ -587,7 +587,7 @@ func TestPoSDownloader(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	payloadStatus := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_SYNCING, payloadStatus.Status)
+	assert.Equal(t, engine.EngineStatus_SYNCING, payloadStatus.Status)
 
 	// Send the missing header
 	b, err := rlp.EncodeToBytes(&eth.BlockHeadersPacket66{
@@ -612,7 +612,7 @@ func TestPoSDownloader(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	// Point forkChoice to the head
-	forkChoiceMessage := engineapi.ForkChoiceMessage{
+	forkChoiceMessage := engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      chain.TopBlock.Hash(),
 		SafeBlockHash:      chain.TopBlock.Hash(),
 		FinalizedBlockHash: chain.TopBlock.Hash(),
@@ -624,7 +624,7 @@ func TestPoSDownloader(t *testing.T) {
 	assert.Equal(t, chain.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 
 	payloadStatus = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus.Status)
 	assert.Equal(t, chain.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 }
 
@@ -658,7 +658,7 @@ func TestPoSSyncWithInvalidHeader(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	payloadStatus1 := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_SYNCING, payloadStatus1.Status)
+	assert.Equal(t, engine.EngineStatus_SYNCING, payloadStatus1.Status)
 
 	// Send the missing headers
 	b, err := rlp.EncodeToBytes(&eth.BlockHeadersPacket66{
@@ -677,7 +677,7 @@ func TestPoSSyncWithInvalidHeader(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	// Point forkChoice to the invalid tip
-	forkChoiceMessage := engineapi.ForkChoiceMessage{
+	forkChoiceMessage := engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      invalidTip.Hash(),
 		SafeBlockHash:      invalidTip.Hash(),
 		FinalizedBlockHash: invalidTip.Hash(),
@@ -748,8 +748,8 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 
 	payloadStatus0 := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus0.Status)
-	forkChoiceMessage := engineapi.ForkChoiceMessage{
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus0.Status)
+	forkChoiceMessage := engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      chain0.TopBlock.Hash(),
 		SafeBlockHash:      chain0.TopBlock.Hash(),
 		FinalizedBlockHash: chain0.TopBlock.Hash(),
@@ -760,7 +760,7 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	assert.Equal(t, chain0.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 	payloadStatus0 = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus0.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus0.Status)
 	assert.Equal(t, chain0.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 
 	//------------------------------------------
@@ -769,8 +769,8 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	require.NoError(err)
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	payloadStatus1 := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus1.Status)
-	forkChoiceMessage = engineapi.ForkChoiceMessage{
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus1.Status)
+	forkChoiceMessage = engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      chain1.TopBlock.Hash(),
 		SafeBlockHash:      chain1.TopBlock.Hash(),
 		FinalizedBlockHash: chain1.TopBlock.Hash(),
@@ -781,7 +781,7 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	assert.Equal(t, chain1.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 	payloadStatus1 = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus1.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus1.Status)
 	assert.Equal(t, chain1.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 
 	//------------------------------------------
@@ -790,8 +790,8 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	require.NoError(err)
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	payloadStatus2 := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus2.Status)
-	forkChoiceMessage = engineapi.ForkChoiceMessage{
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus2.Status)
+	forkChoiceMessage = engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      chain2.TopBlock.Hash(),
 		SafeBlockHash:      chain2.TopBlock.Hash(),
 		FinalizedBlockHash: chain2.TopBlock.Hash(),
@@ -802,7 +802,7 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	assert.Equal(t, chain2.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 	payloadStatus2 = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus2.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus2.Status)
 	assert.Equal(t, chain2.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 
 	//------------------------------------------
@@ -812,14 +812,14 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	require.NoError(err)
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	payloadStatus3 := m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus3.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus3.Status)
 	m.SendPayloadRequest(chain3.TopBlock)
 	err = stages.StageLoopIteration(m.Ctx, m.DB, tx, m.Sync, initialCycle, m.Log, m.BlockReader, nil)
 	require.NoError(err)
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	payloadStatus3 = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus3.Status)
-	forkChoiceMessage = engineapi.ForkChoiceMessage{
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus3.Status)
+	forkChoiceMessage = engine_helpers.ForkChoiceMessage{
 		HeadBlockHash:      chain3.TopBlock.Hash(),
 		SafeBlockHash:      chain3.TopBlock.Hash(),
 		FinalizedBlockHash: chain3.TopBlock.Hash(),
@@ -830,6 +830,6 @@ func TestPOSWrongTrieRootReorgs(t *testing.T) {
 	stages.SendPayloadStatus(m.HeaderDownload(), rawdb.ReadHeadBlockHash(tx), err)
 	assert.Equal(t, chain3.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 	payloadStatus3 = m.ReceivePayloadStatus()
-	assert.Equal(t, remote.EngineStatus_VALID, payloadStatus3.Status)
+	assert.Equal(t, engine.EngineStatus_VALID, payloadStatus3.Status)
 	assert.Equal(t, chain3.TopBlock.Hash(), rawdb.ReadHeadBlockHash(tx))
 }
